gradio_version/requirements_gradio.txt
# Dependências principais
gradio>=4.0.0
streamlit>=1.28.0

# Microsoft Presidio - Core (versões estáveis)
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Processamento de linguagem natural
spacy>=3.5.0
presidio-analyzer[spacy]>=2.2.0

# Processamento de documentos
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21

# Utilitários
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=9.0.0

# Modelos de linguagem (opcionais)
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
groq>=0.4.0

# Análise de texto
nltk>=3.8.1
textblob>=0.17.1

# Processamento de dados
pandas>=1.5.0
numpy>=1.21.0

# Interface e visualização
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilitários de sistema
psutil>=5.9.0
tqdm>=4.65.0